All dependencies for movie DEG is active
Updated the GameObject's for the movie DEG sub-objects to be set active.
However the main GameObject 'Movie_1 [DEG Complex]' is set as inactive
in compliance to the movie player back-end.  However, this compliance is
set-fourth as a legacy mechanism - not the Video Player component that
was newly introduced in Unity3D 5.6.x.